WIP: more support for PIPE

+++ FreeRTOS/Demo/RISC-V_Galois_P1/bsp/plic_driver.c
@@ -9,10 +9,10 @@ static void volatile_memzero(uint8_t *base, unsigned int size);
 
 static void volatile_memzero(uint8_t *base, unsigned int size)
 {
-    volatile uint8_t *ptr;
-    for (ptr = base; ptr < (base + size); ptr++)
-    {
-        *ptr = 0;
+    // XXX: This assumes base and size are word-aligned
+    volatile uint32_t * ptr;
+    for (ptr = base; ptr < (base + size); ptr++) {
+        *ptr = (uint32_t)0;
     }
 }
